Refactor: controller responses
