+++ escout/Controllers/AuthenticationController.cs
@@ -1,6 +1,10 @@
-﻿using escout.Models;
+﻿using System;
+using System.Linq;
+using escout.Helpers;
+using escout.Models;
 using escout.Services;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace escout.Controllers
@@ -12,51 +16,100 @@ namespace escout.Controllers
         private readonly DataContext context;
         public AuthenticationController(DataContext context) => this.context = context;
 
-        /// <summary>
-        /// Generate authentication token.
-        /// </summary>
         [HttpPost]
         [AllowAnonymous]
         [Route("signIn")]
-        public ActionResult<AuthData> SignIn(User userData)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public ActionResult<AuthData> SignIn(User user)
         {
-            using var service = new AuthService(context);
-            var user = service.SignIn(userData);
+            var account = context.users.FirstOrDefault(u => u.username == user.username.ToLower());
 
-            if (user != null)
+            if (account != null)
             {
-                var token = TokenService.GenerateToken(user);
-                return token;
+                if (TokenService.VerifyPassword(user.password, account.password))
+                    return TokenService.GenerateToken(account);
+                else
+                    return BadRequest("Password is wrong.");
             }
-
-            return new NotFoundResult();
+            else
+                return BadRequest("Account does not exist.");
         }
 
-        /// <summary>
-        /// Create user.
-        /// </summary>
         [HttpPost]
         [AllowAnonymous]
         [Route("signUp")]
-        public ActionResult<SvcResult> SignUp(User user)
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public IActionResult SignUp(User user)
         {
-            using var userService = new UserService(context);
-            using var authService = new AuthService(context);
+            if (CheckEmailExist(user.email))
+                return BadRequest("Email in use");
+            if (CheckUsernameExist(user.username))
+                return BadRequest("Username in use");
 
-            if (userService.CheckEmailExist(user.email)) return SvcResult.Set(1, "Email in use");
-            if (userService.CheckUsernameExist(user.username)) return SvcResult.Set(1, "User in use");
-            return authService.SignUp(user) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error while adding user");
+            try
+            {
+                user.accessLevel = 0;
+                user.created = Utils.GetDateTime();
+                user.updated = Utils.GetDateTime();
+                user.username = user.username.ToLower();
+                user.email = user.email.ToLower();
+                user.password = TokenService.HashPassword(user.password);
+                context.users.Add(user);
+                context.SaveChanges();
+
+                if (user.notifications == 1)
+                    _ = NotificationHelper.SendEmail(user.email, "Welcome to eScout", "Welcome to eScout " + user.username);
+
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
         }
 
-        /// <summary>
-        /// Test authentication token.
-        /// </summary>
+        [HttpPost]
+        [AllowAnonymous]
+        [Route("resetPassword")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public IActionResult ResetPassword(User user1)
+        {
+            var user = context.users.FirstOrDefault(u => u.username == user1.username || u.email == user1.email);
+
+            if (user == null)
+                return BadRequest("Account does not exist");
+
+            var generatedPassword = Utils.StringGenerator();
+            user.updated = Utils.GetDateTime();
+            user.password = TokenService.HashPassword(Utils.GenerateSha256String(generatedPassword));
+            context.users.Update(user);
+            context.SaveChanges();
+            _ = NotificationHelper.SendEmail(user.email, "New eScout Password", generatedPassword);
+            return Ok();
+        }
+
         [HttpGet]
         [Authorize]
         [Route("authenticated")]
-        public ActionResult<SvcResult> Authenticated()
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        public IActionResult Authenticated()
         {
-            return SvcResult.Set(0, "Success");
+            return Ok();
+        }
+
+        private bool CheckEmailExist(string email)
+        {
+            var check = context.users.FirstOrDefault(u => u.email == email);
+            return check != null;
+        }
+
+        private bool CheckUsernameExist(string username)
+        {
+            var check = context.users.FirstOrDefault(u => u.username == username);
+            return check != null;
         }
     }
 }
