+++ escout/Controllers/GameController.cs
@@ -1,13 +1,18 @@
-﻿using escout.Models;
+﻿using escout.Helpers;
+using escout.Models;
 using escout.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace escout.Controllers
 {
-    [Route("api/v1")]
+    [Authorize]
     [ApiController]
+    [Route("api/v1")]
     public class GameController : ControllerBase
     {
         private readonly DataContext context;
@@ -17,66 +22,70 @@ namespace escout.Controllers
             this.context = context;
         }
 
-        /// <summary>
-        /// Create game.
-        /// </summary>
         [HttpPost]
-        [Authorize]
         [Route("game")]
         public ActionResult<List<Game>> CreateGame(List<Game> game)
         {
-            using var service = new GameService(context);
-            return service.CreateGame(game);
+            game.ToList().ForEach(g => g.created = Utils.GetDateTime());
+            game.ToList().ForEach(g => g.updated = Utils.GetDateTime());
+            context.games.AddRange(game);
+            context.SaveChanges();
+            return game;
         }
 
-        /// <summary>
-        /// Update game.
-        /// </summary>
         [HttpPut]
-        [Authorize]
         [Route("game")]
-        public ActionResult<SvcResult> UpdateGame(Game game)
+        public IActionResult UpdateGame(Game game)
         {
-            using var service = new GameService(context);
-            return service.UpdateGame(game) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                game.updated = Utils.GetDateTime();
+                context.games.Update(game);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
-        /// <summary>
-        /// Delete game.
-        /// </summary>
         [HttpDelete]
-        [Authorize]
         [Route("game")]
-        public ActionResult<SvcResult> DeleteGame(int id)
+        public IActionResult DeleteGame(int id)
         {
-            using var service = new GameService(context);
-            return service.DeleteGame(id) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                var game = context.games.FirstOrDefault(g => g.id == id);
+                context.games.Remove(game);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
-        /// <summary>
-        /// Get game.
-        /// </summary>
         [HttpGet]
-        [Authorize]
         [Route("game")]
         public ActionResult<Game> GetGame(int id)
         {
-            using var service = new GameService(context);
-            return service.GetGame(id);
+            return context.games.FirstOrDefault(g => g.id == id);
         }
 
-        /// <summary>
-        /// Get games.
-        /// </summary>
         [HttpGet]
-        [Authorize]
         [Route("games")]
         public ActionResult<List<Game>> GetGames(string query)
         {
             try
             {
-                using var service = new GameService(context);
-                return service.GetGames(query);
+                List<Game> games;
+
+                if (string.IsNullOrEmpty(query))
+                    games = context.games.ToList();
+                else
+                {
+                    var criteria = JsonConvert.DeserializeObject<FilterCriteria>(query);
+                    var q = string.Format("SELECT * FROM games WHERE " + criteria.fieldName + " " + criteria.condition + " '" + criteria.value + "';");
+                    games = context.games.FromSqlRaw(q).ToList();
+                }
+
+                return games;
             }
             catch
             {
@@ -85,96 +94,123 @@ namespace escout.Controllers
         }
 
         [HttpPost]
-        [Authorize]
         [Route("gameEvent")]
-        public ActionResult<SvcResult> CreateGameEvent(List<GameEvent> gameEvent)
+        public IActionResult CreateGameEvent(List<GameEvent> gameEvent)
         {
-            var user = User.GetUser(new UserService(context));
-            using var service = new GameService(context);
-            return service.CreateGameEvent(gameEvent, user).Count > 0 ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            gameEvent.ToList().ForEach(g => g.userId = User.GetUser(context).id);
+            gameEvent.ToList().ForEach(g => g.created = Utils.GetDateTime());
+            gameEvent.ToList().ForEach(g => g.updated = Utils.GetDateTime());
+            context.gameEvents.AddRange(gameEvent);
+            context.SaveChanges();
+            return gameEvent.Count > 0 ? Ok() : BadRequest();
         }
 
         [HttpPut]
-        [Authorize]
         [Route("gameEvent")]
-        public ActionResult<SvcResult> UpdateGameEvent(GameEvent gameEvent)
+        public IActionResult UpdateGameEvent(GameEvent gameEvent)
         {
-            using var service = new GameService(context);
-            return service.UpdateGameEvent(gameEvent) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                gameEvent.updated = Utils.GetDateTime();
+                context.gameEvents.Update(gameEvent);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpDelete]
-        [Authorize]
         [Route("gameEvent")]
-        public ActionResult<SvcResult> DeleteGameEvent(int id)
+        public IActionResult DeleteGameEvent(int id)
         {
-            using var service = new GameService(context);
-            return service.DeleteGameEvent(id) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                var gameEvent = context.gameEvents.FirstOrDefault(g => g.id == id);
+                context.gameEvents.Remove(gameEvent);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameEvent")]
         public ActionResult<GameEvent> GetGameEvent(int id)
         {
-            using var service = new GameService(context);
-            return service.GetGameEvent(id);
+            return context.gameEvents.FirstOrDefault(g => g.id == id);
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameEvents")]
         public ActionResult<List<GameEvent>> GetGameEvents(int gameId)
         {
-            using var service = new GameService(context);
-            return service.GetGameEvents(gameId);
+            return context.gameEvents.Where(g => g.gameId == gameId).ToList();
         }
 
         [HttpPost]
-        [Authorize]
         [Route("gameAthlete")]
         public ActionResult<List<GameAthlete>> CreateGameAthlete(List<GameAthlete> gameAthlete)
         {
-            using var service = new GameService(context);
-            return service.CreateGameAthlete(gameAthlete);
+            gameAthlete.ToList().ForEach(g => g.created = Utils.GetDateTime());
+            gameAthlete.ToList().ForEach(g => g.updated = Utils.GetDateTime());
+            context.gameAthletes.AddRange(gameAthlete);
+            context.SaveChanges();
+            return gameAthlete;
         }
 
         [HttpPut]
-        [Authorize]
         [Route("gameAthlete")]
-        public ActionResult<SvcResult> UpdateGameAthlete(GameAthlete gameAthlete)
+        public IActionResult UpdateGameAthlete(GameAthlete gameAthlete)
         {
-            using var service = new GameService(context);
-            return service.UpdateGameAthlete(gameAthlete) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                gameAthlete.updated = Utils.GetDateTime();
+                context.gameAthletes.Update(gameAthlete);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpDelete]
-        [Authorize]
         [Route("gameAthlete")]
-        public ActionResult<SvcResult> DeleteGameAthlete(int id)
+        public IActionResult DeleteGameAthlete(int id)
         {
-            using var service = new GameService(context);
-            return service.DeleteGameAthlete(id) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                var gameAthlete = context.gameAthletes.FirstOrDefault(g => g.id == id);
+                context.gameAthletes.Remove(gameAthlete);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameAthletes")]
         public ActionResult<List<GameAthlete>> GetGameAthletes(int gameId)
         {
-            using var service = new GameService(context);
-            return service.GetGamesAthletes(gameId);
+            return context.gameAthletes.Where(g => g.gameId == gameId).ToList();
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameUser")]
         public ActionResult<List<GameUser>> GetGameUser(string query)
         {
             try
             {
-                using var service = new GameService(context);
-                return service.GetGameUsers(query);
+                List<GameUser> gamesUsers;
+
+                if (string.IsNullOrEmpty(query))
+                    gamesUsers = context.gameUsers.ToList();
+                else
+                {
+                    var criteria = JsonConvert.DeserializeObject<FilterCriteria>(query);
+                    var q = string.Format("SELECT * FROM gameUsers WHERE " + criteria.fieldName + " " + criteria.condition + " '" + criteria.value + "';");
+                    gamesUsers = context.gameUsers.FromSqlRaw(q).ToList();
+                }
+
+                return gamesUsers;
             }
             catch
             {
@@ -183,57 +219,109 @@ namespace escout.Controllers
         }
 
         [HttpPost]
-        [Authorize]
         [Route("gameUser")]
         public ActionResult<List<GameUser>> CreateGameUser(List<GameUser> gameUsers)
         {
-            using var service = new GameService(context);
-            return service.CreateGameUser(gameUsers);
+            foreach (var gameUser in gameUsers)
+            {
+                var obj = context.gameUsers.Where(c => c.gameId == gameUser.gameId && c.userId == gameUser.userId && c.athleteId == gameUser.athleteId).ToList();
+                if (obj.Count != 0)
+                    gameUsers.Remove(gameUser);
+            }
+
+            gameUsers.ToList().ForEach(g => g.created = Utils.GetDateTime());
+            gameUsers.ToList().ForEach(g => g.updated = Utils.GetDateTime());
+            context.gameUsers.AddRange(gameUsers);
+            context.SaveChanges();
+            return gameUsers;
         }
 
         [HttpPut]
-        [Authorize]
         [Route("gameUser")]
-        public ActionResult<SvcResult> UpdateGameUser(GameUser gameUser)
+        public IActionResult UpdateGameUser(GameUser gameUser)
         {
-            using var service = new GameService(context);
-            return service.UpdateGameUser(gameUser) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                gameUser.updated = Utils.GetDateTime();
+                context.gameUsers.Update(gameUser);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpDelete]
-        [Authorize]
         [Route("gameUser")]
-        public ActionResult<SvcResult> DeleteGameUser(GameUser gameUser)
+        public IActionResult DeleteGameUser(GameUser gameUser)
         {
-            using var service = new GameService(context);
-            return service.DeleteGameUser(gameUser) ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+            try
+            {
+                context.gameUsers.Remove(gameUser);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
         }
 
         [HttpGet]
-        [Authorize]
         [Route("athleteGameEvents")]
         public ActionResult<List<GameEvent>> AthleteGameEvents(int athleteId, int gameId)
         {
-            using var service = new GameService(context);
-            return service.AthleteGameEvents(athleteId, gameId);
+            return context.gameEvents.Where(g => g.athleteId == athleteId && g.gameId == gameId).ToList();
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameData")]
         public ActionResult<GameData> GetGameData(int gameId)
         {
-            using var service = new GameService(context);
-            return service.GetGameData(gameId);
+            var gameData = new GameData();
+            gameData.game = context.games.FirstOrDefault(g => g.id == gameId);
+            gameData.clubs = context.clubs.Where(t => t.id == gameData.game.homeId || t.id == gameData.game.visitorId).ToList();
+            gameData.athletes = context.athletes.Where(t => t.clubId == gameData.game.homeId || t.clubId == gameData.game.visitorId).ToList();
+
+            if (gameData.game.competitionId != null)
+            {
+                gameData.competition = context.competitions.FirstOrDefault(t => t.id == gameData.game.competitionId);
+                gameData.sport = context.sports.FirstOrDefault(t => t.id == gameData.competition.sportId);
+            }
+            else
+                gameData.sport = context.sports.FirstOrDefault(t => t.name == "Soccer");
+
+            gameData.events = context.events.Where(t => t.sportId == gameData.sport.id).ToList();
+            gameData.gameEvents = context.gameEvents.Where(t => t.gameId == gameId).ToList();
+
+            return gameData;
         }
 
         [HttpGet]
-        [Authorize]
         [Route("gameStatistics")]
         public ActionResult<List<ClubStats>> GetGameStatistics(int gameId)
         {
-            using var service = new GameService(context);
-            return service.GetGameStatistics(gameId);
+            var stats = new List<ClubStats>();
+
+            var gameEvents = context.gameEvents.Where(x => x.gameId == gameId).ToList();
+            var uniqueClubs = gameEvents.Select(x => x.clubId).Distinct().ToList();
+
+            foreach (var i in uniqueClubs)
+            {
+                if (i != null)
+                {
+                    var clubEvents = gameEvents.Where(x => x.clubId == i).ToList();
+                    foreach (Event e in context.events)
+                    {
+                        var evt = clubEvents.Where(x => x.eventId == e.id).ToList();
+                        var stat = new ClubStats
+                        {
+                            Count = evt.Count(),
+                            ClubId = int.Parse(i.ToString()),
+                            EventId = e.id
+                        };
+                        stats.Add(stat);
+                    }
+                }
+            }
+
+            return stats;
         }
     }
 }
