+++ escout/Controllers/UserController.cs
@@ -1,110 +1,84 @@
-﻿using escout.Models;
+﻿using escout.Helpers;
+using escout.Models;
 using escout.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace escout.Controllers
 {
-    [Route("api/v1")]
+    [Authorize]
     [ApiController]
+    [Route("api/v1")]
     public class UserController : ControllerBase
     {
         private readonly DataContext context;
         public UserController(DataContext context) => this.context = context;
 
-        /// <summary>
-        /// Reset user password.
-        /// </summary>
         [HttpPost]
-        [Route("resetPassword")]
-        public ActionResult<SvcResult> ResetPassword(User user)
-        {
-            using var service = new UserService(context);
-            var result = service.ResetPassword(user.username, user.email);
-            return result ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
-        }
-
-        /// <summary>
-        /// Change user password.
-        /// </summary>
-        [HttpPost]
-        [Authorize]
         [Route("changePassword")]
-        public ActionResult<SvcResult> ChangePassword(string newPassword)
+        public IActionResult ChangePassword(string newPassword)
         {
-            var user = User.GetUser(new UserService(context));
+            var user = User.GetUser(context);
             if (user == null)
                 return new NotFoundResult();
 
-            using var userService = new UserService(context);
-            var result = userService.ChangePassword(user, newPassword);
-            return result ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
-        }
-
-        /// <summary>
-        /// Update user.
-        /// </summary>
-        [HttpPut]
-        [Authorize]
-        [Route("user")]
-        public ActionResult<SvcResult> UpdateUser(User user)
-        {
-            var currentUser = User.GetUser(new UserService(context));
-            if (currentUser.accessLevel != 0)
-                user.id = currentUser.id;
-
-            using var userService = new UserService(context);
-            var result = userService.UpdateUser(user);
-            return result ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
-        }
-
-        /// <summary>
-        /// Delete user.
-        /// </summary>
-        [HttpDelete]
-        [Authorize]
-        [Route("user")]
-        public ActionResult<SvcResult> RemoveUser(User user)
-        {
-            if (user.accessLevel == 0)
+            try
             {
-                using var service = new UserService(context);
-                var result = service.RemoveUser(user);
-                return result ? SvcResult.Set(0, "Success") : SvcResult.Set(1, "Error");
+                user.password = BCrypt.Net.BCrypt.HashPassword(newPassword);
+                user.updated = Utils.GetDateTime();
+                context.users.Update(user);
+                context.SaveChanges();
+                return Ok();
             }
-            else
-                return new UnauthorizedResult();
+            catch { return BadRequest(); }
         }
 
-        /// <summary>
-        /// Get user.
-        /// </summary>
+        [HttpPut]
+        [Route("user")]
+        public IActionResult UpdateUser(User user)
+        {
+            try
+            {
+                user.updated = Utils.GetDateTime();
+                context.users.Update(user);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
+        }
+
+        [HttpDelete]
+        [Route("user")]
+        public IActionResult DeleteUser(User user)
+        {
+            try
+            {
+                context.users.Remove(user);
+                context.SaveChanges();
+                return Ok();
+            }
+            catch { return BadRequest(); }
+        }
+
         [HttpGet]
         [Route("user")]
-        [Authorize]
         public ActionResult<User> GetUser()
         {
-            return User.GetUser(new UserService(context));
+            return User.GetUser(context);
         }
 
-        /// <summary>
-        /// Get users.
-        /// </summary>
         [HttpGet]
-        [Authorize]
         [Route("users")]
         public ActionResult<List<User>> GetUsers(string query)
         {
-            var user = User.GetUser(new UserService(context));
+            var user = User.GetUser(context);
 
             if (user == null)
                 return new NotFoundResult();
             else if (user.accessLevel == 0)
-            {
-                using var service = new UserService(context);
-                return service.GetUsers();
-            }
+                return context.users.ToList();
             else
                 return new UnauthorizedResult();
         }
