+++ escoutTests/Controllers/ImageControllerTests.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using escout.Models;
 using escoutTests.Resources;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace escout.Controllers.Tests
@@ -42,8 +43,8 @@ namespace escout.Controllers.Tests
             image.imageUrl = "test image";
             var result = controller.UpdateImage(image);
 
-            Assert.AreEqual(0, result.Value.errorCode);
             Assert.AreEqual(image.imageUrl, context.images.First().imageUrl);
+            Assert.AreEqual(200, ((StatusCodeResult)result).StatusCode);
         }
 
         [TestMethod]
@@ -52,8 +53,8 @@ namespace escout.Controllers.Tests
             TestUtils.AddImageToContext(context);
             var result = controller.DeleteImage(context.images.First().id);
 
-            Assert.AreEqual(0, result.Value.errorCode);
             Assert.AreEqual(0, context.images.Count());
+            Assert.AreEqual(200, ((StatusCodeResult)result).StatusCode);
         }
 
         [TestMethod]
