+++ escoutTests/Controllers/UserControllerTests.cs
@@ -1,5 +1,6 @@
 ﻿using escout.Models;
 using escoutTests.Resources;
+using Microsoft.AspNetCore.Mvc;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Linq;
 
@@ -46,18 +47,18 @@ namespace escout.Controllers.Tests
             user.username = "testuser";
             var result = controller.UpdateUser(user);
 
-            Assert.AreEqual(0, result.Value.errorCode);
             Assert.AreEqual(user.username, context.users.First().username);
+            Assert.AreEqual(200, ((StatusCodeResult)result).StatusCode);
         }
 
         [TestMethod]
         public void RemoveUserTest()
         {
             var user = TestUtils.AddUserToContext(context);
-            var result = controller.RemoveUser(user);
+            var result = controller.DeleteUser(user);
 
-            Assert.AreEqual(0, result.Value.errorCode);
             Assert.AreEqual(0, context.users.Count());
+            Assert.AreEqual(200, ((StatusCodeResult)result).StatusCode);
         }
 
         [Ignore]
@@ -68,16 +69,18 @@ namespace escout.Controllers.Tests
             var result = controller.GetUser();
 
             Assert.AreEqual(user.username, result.Value.username);
+            Assert.AreEqual(200, ((StatusCodeResult)result.Result).StatusCode);
         }
 
         [Ignore]
         [TestMethod]
         public void GetUsersTest()
         {
-            var user = TestUtils.AddUserToContext(context);
+            TestUtils.AddUserToContext(context);
             var result = controller.GetUsers(string.Empty);
 
             Assert.AreEqual(1, result.Value.Count);
+            Assert.AreEqual(200, ((StatusCodeResult)result.Result).StatusCode);
         }
     }
 }
